Move tests from dev
